src/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
attrs==24.2.0
cffi==1.17.1
discord==2.3.2
discord.py==2.4.0
dpkt==1.9.8
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
nfstream==6.5.3
numpy==2.1.3
pandas==2.2.3
propcache==0.2.0
psutil==6.1.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tzdata==2024.2
yarl==1.17.1